backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
bcrypt>=3.2.0
PyJWT>=2.3.0 